Error check for undefined access token
After a lot of troubleshooting, I realized I'd made a very dumb easy typo when I updated my config for .12. Adding this as a way to catch this condition so you don't just mysteriously fail with 403, etc. 

Note, I'm not a JS expert, but this seemed to work. Figured I'd try to do the PR vs just complaining in an issue. 😄